# GeLoRa/backend/docs/swagger/swagger.yaml
definitions:
  adaptersinboundhttpdto.CommonErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  adaptersinboundhttpdto.GetUserInfoByIdResponse:
    properties:
      bio:
        example: Hello, I am John.
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      role:
        allOf:
        - $ref: '#/definitions/domainmodel.UserRole'
        example: client
      updated_at:
        example: "2024-06-01T00:00:00Z"
        type: string
      username:
        example: john_doe_99
        type: string
    type: object
...
  adaptersinboundhttpdto.GetUsersListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/adaptersinboundhttpdto.GetUsersListData'
        type: array
      pagination:
        $ref: '#/definitions/adaptersinboundhttpdto.CommonPaginationMeta'
    type: object
...
  adaptersinboundhttpdto.PatchUserInfoByIdRequest:
    properties:
      bio:
        example: Hello, I am John.
        type: string
      name:
        example: John Doe
        type: string
    type: object
...
  adaptersinboundhttpdto.PostRefreshResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  adaptersinboundhttpdto.PostSignInRequest:
    properties:
      password:
        example: secret123
        type: string
      username:
        example: john_doe_99
        type: string
    type: object
  adaptersinboundhttpdto.PostSignInResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  adaptersinboundhttpdto.PostSignUpRequest:
    properties:
      name:
        example: John Doe
        type: string
      password:
        example: secret123
        type: string
      username:
        example: john_doe_99
        type: string
    type: object
  adaptersinboundhttpdto.PostSignUpResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  adaptersinboundhttpdto.PutUserPasswordByIdRequest:
    properties:
      new_password:
        example: newsecret456
        type: string
    type: object
...
  domainmodel.UserRole:
    enum:
    - super
    - admin
    - client
    type: string
    x-enum-varnames:
    - UserRoleSuper
    - UserRoleAdmin
    - UserRoleClient
info:
  contact: {}
  description: The core API documentation for GeLoRa backend endpoints.
  title: GeLoRa API
  version: "1.0"
paths:
  /auth/refresh:
    post:
      description: Issue a new access and refresh token pair using a valid refresh
        token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.PostRefreshResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Refresh tokens
      tags:
      - Auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return access and refresh tokens
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/adaptersinboundhttpdto.PostSignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.PostSignInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      summary: Sign in
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Register a new user and return access and refresh tokens
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/adaptersinboundhttpdto.PostSignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.PostSignUpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      summary: Sign up
      tags:
      - Auth
...
  /users:
    get:
      description: Return a paginated list of users with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Cursor ID for pagination
        in: query
        name: cursor_id
        type: integer
      - description: Search by name or username
        in: query
        name: search
        type: string
      - description: Filter by role (super, admin, client)
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.GetUsersListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
  /users/{id}:
    delete:
      description: Permanently remove a user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user by ID
      tags:
      - Users
    get:
      description: Return the profile of a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.GetUserInfoByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update the name and/or bio of a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/adaptersinboundhttpdto.PatchUserInfoByIdRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user by ID
      tags:
      - Users
  /users/{id}/password:
    put:
      consumes:
      - application/json
      description: Directly set a new password for a specific user (admin, no old
        password required)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/adaptersinboundhttpdto.PutUserPasswordByIdRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Override user password by ID
      tags:
      - Users
  /users/{id}/role:
    put:
      consumes:
      - application/json
      description: Assign a role to a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/adaptersinboundhttpdto.PutUserRoleByIdRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Set user role by ID
      tags:
      - Users
  /users/me:
    get:
      description: Return the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.GetUserInfoMeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update the name and/or bio of the currently authenticated user
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/adaptersinboundhttpdto.PatchUserInfoMeRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Update my profile
      tags:
      - Users
  /users/me/password:
    put:
      consumes:
      - application/json
      description: Change the password of the currently authenticated user (requires
        old password)
      parameters:
      - description: Old and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/adaptersinboundhttpdto.PutUserPasswordMeRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Change my password
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"